requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
transformers
datasets
sentence-transformers
faiss-cpu
langchain
python-dotenv
